[MIG][15.0] Module vault and vault_share
