+++ vault/__manifest__.py
@@ -5,7 +5,7 @@
     "name": "Vault",
     "summary": "Password vault integration in Odoo",
     "license": "AGPL-3",
-    "version": "14.0.1.8.0",
+    "version": "15.0.1.6.1",
     "website": "https://github.com/OCA/server-auth",
     "application": True,
     "author": "initOS GmbH, Odoo Community Association (OCA)",
@@ -14,7 +14,6 @@
     "data": [
         "security/ir.model.access.csv",
         "security/ir_rule.xml",
-        "views/assets.xml",
         "views/res_config_settings_views.xml",
         "views/res_users_views.xml",
         "views/vault_entry_views.xml",
@@ -31,5 +30,24 @@
         "wizards/vault_send_wizard.xml",
         "wizards/vault_store_wizard.xml",
     ],
-    "qweb": ["static/src/xml/templates.xml"],
+    "assets": {
+        "vault.assets_frontend": [
+            "vault/static/src/common/*.js",
+            "vault/static/src/frontend/*.js",
+        ],
+        "web.assets_backend": [
+            "vault/static/lib/**/*.min.js",
+            "vault/static/src/common/*.js",
+            "vault/static/src/backend/*.scss",
+            "vault/static/src/backend/*.js",
+            "vault/static/src/legacy/vault_controller.js",
+            "vault/static/src/legacy/vault_widget.js",
+        ],
+        "web.assets_qweb": [
+            "vault/static/src/**/*.xml",
+        ],
+        "web.tests_assets": [
+            "vault/static/tests/**/*.js",
+        ],
+    },
 }
