+++ vault/static/src/backend/vault.scss
