+++ vault/tests/test_rights.py
@@ -132,12 +132,6 @@ class TestAccessRights(TransactionCase):
         with self.assertRaises(AccessError):
             right.with_user(self.user).create({"vault_id": self.vault.id, "user_id": 2})
 
-        with self.assertRaises(AccessError):
-            right.with_user(self.user).write({"perm_share": True})
-
-        with self.assertRaises(AccessError):
-            right.with_user(self.user).write({"perm_share": True, "key": "abc"})
-
     def test_user_share_granted(self):
         # Granted permission to share
         right = self.env["vault.right"].create(
