+++ vault/views/vault_entry_views.xml
@@ -165,8 +165,18 @@
                     context="{'group_by': 'vault_id'}"
                 />
                 <searchpanel>
-                    <field name="vault_id" string="Vaults" enable_counters="1" />
-                    <field name="parent_id" string="Entries" enable_counters="1" />
+                    <field
+                        name="vault_id"
+                        string="Vaults"
+                        enable_counters="1"
+                        limit="0"
+                    />
+                    <field
+                        name="parent_id"
+                        string="Entries"
+                        enable_counters="1"
+                        limit="0"
+                    />
                 </searchpanel>
             </search>
         </field>
